+++ backend/app/models/api/rolls.py
@@ -6,7 +6,7 @@ class RollCreate(SQLModel):
     roll: int
 
 
-class SessionRollStats(SQLModel):
+class EpisodeRollStats(SQLModel):
     campaign_id: int
     session_id: int
     rolls: list[int]
@@ -23,4 +23,4 @@ class CampaignRollStats(SQLModel):
 
 class RollMutationResponse(SQLModel):
     campaign_stats: CampaignRollStats
-    session_stats: SessionRollStats
+    session_stats: EpisodeRollStats